requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
langchain
langchain-openai
openai
pydantic
python-dotenv
requests
chromadb